@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
 dependencies:
   flutter_localizations:
@@ -35,7 +35,6 @@ dependencies:
   equatable: ^1.2.4
   
   rxdart: ^0.24.1
-  # flutter_barcode_scanner: ^1.0.2
   qr_code_scanner: ^0.3.0
 
 
