@@ -1,20 +1,37 @@
 import 'dart:async';
-
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
+import 'package:rxdart/rxdart.dart';
+
+import '../../repositories/user_repository.dart';
 
 part 'restore_wallet_event.dart';
 part 'restore_wallet_state.dart';
 
 class RestoreWalletBloc extends Bloc<RestoreWalletEvent, RestoreWalletState> {
-  RestoreWalletBloc() : super(RestoreWalletInitial());
+  UserRepository _repo;
+  RestoreWalletBloc(this._repo) : super(RestoreWalletInitial());
+
+  @override
+  Stream<Transition<RestoreWalletEvent, RestoreWalletState>> transformEvents(
+      Stream<RestoreWalletEvent> events, transitionFn) {
+    return events
+        .throttleTime(const Duration(milliseconds: 500))
+        .switchMap((transitionFn));
+  }
 
   @override
   Stream<RestoreWalletState> mapEventToState(
     RestoreWalletEvent event,
   ) async* {
     if (event is GetPaperWallet) {
-      yield PaperWalletSuccess(event.paperWallet);
+      if (state is RestoreWalletInitial || state is PaperWalletFail) {
+        if (_repo.validPaperWallet(event.paperWallet)) {
+          yield PaperWalletSuccess(event.paperWallet);
+        } else {
+          yield PaperWalletFail();
+        }
+      }
     }
   }
 }
